ml/requirements.txt
numpy>=1.24,<3.0
pandas>=2.1
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.3
requests>=2.31
